+++ Renderer/Include/Renderer/TextureBuffer.h
@@ -10,20 +10,20 @@ enum struct TextureFormat {
 
 class TextureBuffer : public RenderResource
 {
-	glm::vec2 m_dimensions;
+	glm::ivec2 m_dimensions;
 	TextureFormat m_format;
 	std::string m_filename;
 	bool fileTexture;
 
 public:
-	TextureBuffer(glm::vec2 dimensions, TextureFormat format);
-	TextureBuffer(glm::vec2 dimensions, TextureFormat format, std::string filename);
+	TextureBuffer(glm::ivec2 dimensions, TextureFormat format);
+	TextureBuffer(glm::ivec2 dimensions, TextureFormat format, std::string filename);
 	~TextureBuffer();
 
 	const std::string& filename() const;
 	bool isFileTexture() const;
 	void* data() const;
 	TextureFormat format() const;
-	glm::vec2 dimensions() const;
+	glm::ivec2 dimensions() const;
 };
